+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
 
-> **Status: Available in ETLantic 0.35.0.**
+> **Status: Available in ETLantic 0.36.0.**
 
 --8<-- "../CHANGELOG.md"